feat: ResultCategoryDetailToggle component
